+++ docs/CHANGELOG.md
@@ -5,6 +5,7 @@ Things under "main" are not released yet.
 ## main
 
 - Remove ability to analyze flow types with `==` to a literal that is not the only value of that type, e.g. `x == 3`. This still allows for flow typing information for cases like `x == null` and `x != null`.
+- Improve correctness of flow typing in other cases.
 
 ## v0.7.7